+++ config/settings.yaml
@@ -2,4 +2,6 @@
 WHITE_LIST:
   SHEET: 凸管理
   CELLS: A2:A100
-#
+
+# ウェルカムメッセージを飛ばすサーバ
+WELCOME_SERVER: べろばあ
