+++ package.json
@@ -19,7 +19,9 @@
   },
   "homepage": "https://github.com/smicle/convex-management#readme",
   "dependencies": {
+    "@types/ws": "^7.2.5",
     "const-settings": "^1.0.0",
+    "discord.js": "^12.2.0",
     "google-spreadsheet-as-promised": "^0.2.12",
     "throw-env": "^1.0.1",
     "typescript": "^3.9.5"
