+++ tsconfig.json
@@ -14,5 +14,5 @@
     "allowJs": true,
     "outDir": "./dist/"
   },
-  "files": ["src/index.ts"]
+  "files": ["src/index.ts", "src/spreadsheet.ts"]
 }
